match: Reject non-const internal functions [PR117260]
When internal functions support was added to match (r6-4979-gc9e926ce2bdc8b),
the check for ECF_CONST was the builtin function side. Though before r15-4503-g8d6d6d537fdc,
there was no use of maybe_push_res_to_seq with non-const internal functions so the check
would not make a difference.

This adds the check for internal functions just as there is a check for builtins.

Note I didn't add a testcase because there was no non-const internal function
which could be used on x86_64 in a decent manor.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	PR tree-optimization/117260
	* gimple-match-exports.cc (maybe_push_res_to_seq): Reject non-const
	internal functions.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>