ada: Further work in semantic analysis of iterated component associations
This finishes up the transition to preanalysis of a copy of the expression
for iterated component associations in all contexts, thus voiding the need
to clean things up afterward.

However, this requires a larger cleanup in semantics analysis of aggregates,
in particular for others choices, which are currently skipped in Sem_Aggr,
with Exp_Aggr trying to patch things up afterward but leaving some legality
loopholes in the end.  That's why this makes sure that all the expressions
appearing in aggregates are either analyzed or preanalyzed by Sem_Aggr, as
documented in the spec of Sem, modulo the copy in an iteration context.

gcc/ada/ChangeLog:

	* exp_aggr.adb (Build_Array_Aggr_Code): Remove obsolete comment.
	(Convert_To_Positional): Remove Ctyp local variable.
	(Is_Static_Element): Remove Dims parameter and do not preanalyze the
	expression there.
	(Expand_Array_Aggregate): Make Ctyp a constant.
	(Compute_Others_Present): Do not preanalyze the expression there.
	* sem_aggr.adb (Resolve_Array_Aggregate): New Ctyp constant.  Use it
	throughout the procedure to denote the component type.
	(Resolve_Aggr_Expr): Always preanalyze a copy of the expression in
	an iteration context.  Preanalyze it directly when the expander is
	active and the choice may cover multiple components.  Otherwise,
	fully analyze it.
	Do not reanalyze an iterated component association with an others
	choice either when there are positional components.
	(Resolve_Iterated_Component_Association): Do not remove references
	from the expression after invoking Resolve_Aggr_Expr on it.