[target/117316] Fix initializer for riscv code alignment handling
The construct used for initializing the code alignments in a recent change is
causing bootstrap problems on riscv64 as seen in the referenced bugzilla.

This patch adjusts the initializer by pushing the NULL down into each uarch
clause.  Bootstrapped on riscv64, regression test in flight, but given
bootstrap is broken it seemed advisable to move this forward now.

I'm so much looking forward to the day when we have performant hardware for
bootstrap testing...  Sigh.

Anyway, bootstrapped and installing on the trunk.

	PR target/117316
gcc/
	* config/riscv/riscv.cc (riscv_tune_param): Drop initializer.
	(*_tune_info): Add initializers for code alignments.