c++, contracts: Fix reporting of 'observe' contracts.
The c++2a implementation considers both terminated and continued
contracts as 'checked' (which is, of course, true).  However it means
that contract_checked_p() is not usable to differentiate enforce and
observe.  For P2900 contracts we need to check the contract condition
directly.

Fixed by checking the contract conditions directly.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>