Add block parameter to begin_block debug hook
Add a parameter to the begin_block debug hook that is a pointer to the
tree_node of the block in question. CodeView needs this as it records
line numbers of inlined functions in a different manner, so we need to
be able to tell if the block is actually the start of an inlined
function.

gcc/
	* debug.cc (do_nothing_debug_hooks): Change begin_block
	function pointer.
	(debug_nothing_int_int_tree): New function.
	* debug.h (struct gcc_debug_hooks): Add tree parameter to begin_block.
	(debug_nothing_int_int_tree): Add declaration.
	* dwarf2out.cc (dwarf2out_begin_block): Add tree parameter.
	(dwarf2_lineno_debug_hooks): Use new dummy function for begin_block.
	* final.cc (final_scan_insn_1): Pass insn block through to
	debug_hooks->begin_block.
	* vmsdbgout.cc (vmsdbgout_begin_block): Add tree parameter.