aarch64: Use canonicalize_comparison in ccmp expansion [PR117346]
While testing the patch for PR 85605 on aarch64, it was noticed that
imm_choice_comparison.c test failed. This was because canonicalize_comparison
was not being called in the ccmp case. This can be noticed without the patch
for PR 85605 as evidence of the new testcase.

Bootstrapped and tested on aarch64-linux-gnu.

	PR target/117346

gcc/ChangeLog:

	* config/aarch64/aarch64.cc (aarch64_gen_ccmp_first): Call
	canonicalize_comparison before figuring out the cmp_mode/cc_mode.
	(aarch64_gen_ccmp_next): Likewise.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/imm_choice_comparison-1.c: New test.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>